+++ src/main/java/com/project/board/controller/BoardController2.java
@@ -89,7 +89,7 @@ public class BoardController2 {
 			model.addAttribute("originalFileName",originalFileName);
 			
 	        s.addBoard2(new BoardVO(0, title, contents, file.getOriginalFilename(), hits,writedate,memId));
-	        return "board/boardinformation";
+	        return "redirect:./boardinformation";
 
 	    
 	    }
