+++ src/main/resources/static/css/admin.css
@@ -1,3 +1,101 @@
+.card {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    word-wrap: break-word;
+    background-color: #fff;
+    background-clip: border-box;
+    border: 1px solid #e3e3e3;
+    border-radius: 20px;
+    margin-top: 10px;
+    margin-bottom: 20px;
+}
+
+
+.card-body{
+    padding: 1.5rem 1.5rem;
+    flex: 1 1 auto;
+}
+
+.card-title{
+    color: #010101;
+    margin-bottom: 1.2rem;
+    text-transform: capitalize;
+    font-size: 1.125rem;
+    font-weight: 600;
+}
+
+.card-description{
+    margin-bottom: 0.875rem;
+    font-weight: 400;
+    color: #76838f;
+    font-size: 1rem;
+    line-height: 1.3rem;
+    text-decoration: none;
+}
+
+.card-description a:hover{
+    color: black;
+    font-weight: bold;
+}
+
+.table-responsive{
+    overflow-x: auto;
+}
+
+.table-responsive > .table{
+    caption-side: bottom;
+    border-collapse: collapse;
+    --bs-table-bg: transparent;
+    --bs-table-accent-bg: transparent;
+    --bs-table-striped-color: #212529;
+    --bs-table-striped-bg: rgba(0, 0, 0, 0.05);
+    --bs-table-active-color: #212529;
+    --bs-table-active-bg: rgba(0, 0, 0, 0.1);
+    --bs-table-hover-color: #212529;
+    --bs-table-hover-bg: #eaeaf1;
+    width: 100%;
+    margin-bottom: 1rem;
+    color: #212529;
+    vertical-align: top;
+    border-color: #dee2e6;
+    margin-bottom: 0;
+}
+
+.table-responsive > .table > thead{
+    vertical-align: bottom;
+}
+
+.table-responsive > .table > thead, tbody, tfoot, tr, td, th{
+    border-color: inherit;
+    border-style: solid;
+    border-width: 0;
+}
+
+.table-responsive > .table thead th{
+    border-top: 0;
+    border-bottom-width: 1px;
+    font-weight: 600;
+    font-size: 0.875rem;
+}
+
+.table-responsive > .table th, .table-responsive > .table td{
+    vertical-align: middle;
+    line-height: 1;
+    white-space: nowrap;
+    padding: 1.125rem 1.375rem;
+}
+
+.table-responsive > .table > tbody{
+    vertical-align: inherit;
+}
+
+.table-responsive > .table button{
+    width: 60px;
+    height: 40px;
+}
+
+
 
 .adminCRUD {
     padding: 20px;
